Doi kieu du lieu thanh size_t va xoa n

+++ SOURCE/Algorithms/source/shakerSort.cpp
@@ -1,14 +1,14 @@
 #include "../header/shakerSort.hpp"
 
 template<class T>
-void shakerSort(std::vector<T>& array, long long &count_comparison) 
+void shakerSort(std::vector<T>& array, size_t& count_comparison) 
 {
     //cài đặt hàm shakerSort
     int start = 0;                          //biến giữ vị trí biên trái
     int End = array.size() - 1;                        //biến giữ vị trí biên phải
     while (++count_comparison && start <= End)                     //Tương tự với Bubble Sort nhưng chạy từ cả 2 phía chưa được sắp xếp
     {
-        for (int i = start; ++count_comparison && i < End; i++)    //Chạy từ bên trái đẩy phần tử lớn nhất lên trên cùng của vị trí chưa được sắp xếp
+        for (size_t i = start; ++count_comparison && i < End; i++)    //Chạy từ bên trái đẩy phần tử lớn nhất lên trên cùng của vị trí chưa được sắp xếp
         {
             if (++count_comparison && array[i] > array[i + 1])
             {
@@ -34,7 +34,7 @@ void shakerSort(std::vector<T>& array) {
     int End = array.size() - 1;                        //Biến giữ vị trí biên phải
     while (start <= End)                     //Tương tự với Bubble Sort nhưng chạy từ cả 2 phía chưa được sắp xếp
     {
-        for (int i = start; i < End; i++)    //Chạy từ bên trái đẩy phần tử lớn nhất lên trên cùng của vị trí chưa được sắp xếp
+        for (size_t i = start; i < End; i++)    //Chạy từ bên trái đẩy phần tử lớn nhất lên trên cùng của vị trí chưa được sắp xếp
         {
             if (array[i] > array[i + 1])
             {
@@ -62,8 +62,8 @@ template void shakerSort(std::vector<float>&);
 template void shakerSort(std::vector<char>&);
 
 //shakerSort with compare
-template void shakerSort(std::vector<int>&, long long&);
-template void shakerSort(std::vector<long long>&, long long&);
-template void shakerSort(std::vector<double>&, long long&);
-template void shakerSort(std::vector<float>&, long long&);
-template void shakerSort(std::vector<char>&, long long&);
+template void shakerSort(std::vector<int>&, size_t&);
+template void shakerSort(std::vector<long long>&, size_t&);
+template void shakerSort(std::vector<double>&, size_t&);
+template void shakerSort(std::vector<float>&, size_t&);
+template void shakerSort(std::vector<char>&, size_t&);
